[#72] add earn money challenge for testing
