+++ Challenger/ChallengerConfig.cs
@@ -3,4 +3,5 @@
 internal record ChallengerConfig {
 
     public SButton ButtonOpenMenu { get; set; } = SButton.K;
+    public bool DisplayEarnMoneyChallenge { get; set; } = false;
 }
