+++ Challenger/i18n/de.json
@@ -17,6 +17,8 @@
   "BreweryChallenge.RenameRiceJuice": "Reisbier",
   "BreweryChallenge.IncludeBeerVegetablesOnly": "Nur Samen für Bier können gekauft werden",
  
+  "EarnMoneyChallenge": "Kapitalist",
+  
   "HermitChallenge": "Einsiedler",
   "HermitChallenge.CanOnlyLeaveOnSunday": "Du kannst die Farm nur Sonntag verlassen",
