+++ Challenger/i18n/default.json
@@ -17,6 +17,8 @@
   "BreweryChallenge.RenameRiceJuice": "Rice Beer",
   "BreweryChallenge.IncludeBeerVegetablesOnly": "Can only buy seeds for plants that make beer",
 
+  "EarnMoneyChallenge": "Capitalist",
+  
   "HermitChallenge": "Hermit",
   "HermitChallenge.CanOnlyLeaveOnSunday": "You can only leave the farm on Sunday",
